implement and use DELETE and GET.
