+++ KiiThingSDK/Makefile
@@ -7,7 +7,7 @@ build_jansson:
 	$(MAKE) -C jansson
 
 build: build_jansson
-	gcc -shared -fPIC kii_cloud.c kii_custom.c kii_prv_utils.c -L jansson -I curl -I jansson -l jansson -l curl -o libkii.so
+	gcc -shared -fPIC kii_cloud.c kii_custom.c kii_prv_utils.c kii_prv_http_curl.c -L jansson -I curl -I jansson -l jansson -l curl -l ssl -l crypto -o libkii.so
 
 ARCHIVE_FILES = kii_cloud.h kii_cloud.c kii_custom.h kii_custom.c kii_prv_utils.h kii_prv_utils.c kii_prv_types.h
 archive: clean
